backend/pyproject.toml
[project]
name = "aoriarh-backend"
version = "0.1.0"
description = "AORIA RH - Assistant juridique RH par IA"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "email-validator>=2.2.0",
    "llama-index-core>=0.12.0",
    "llama-index-vector-stores-qdrant>=0.4.0",
    "llama-index-embeddings-voyageai>=0.3.0",
    "llama-index-llms-openai>=0.3.0",
    "qdrant-client>=1.13.0",
    "boto3>=1.36.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "stripe>=11.0.0",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "pymupdf4llm>=0.0.17",
    "python-docx>=1.1.0",
    "fastembed>=0.5.0",
    "gunicorn>=22.0.0",
    "slowapi>=0.1.9",
    "arq>=0.26.0",
    "redis>=5.0.0",
    "structlog>=24.0.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"app/services/email/templates.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
